+++ hydromt_delft3dfm/components/mdu.py
@@ -1,5 +1,6 @@
-"""Custom Delft3D-FM config component."""
+"""Custom Delft3D FM config component."""
 
+import datetime as dt
 import logging
 import os
 from os.path import dirname, join
@@ -7,6 +8,7 @@ from pathlib import Path
 from typing import cast
 
 from hydrolib.core.dflowfm import FMModel
+from hydrolib.core.dflowfm.ini.util import validate_datetime_string
 from hydromt import hydromt_step
 from hydromt.model import Model
 from hydromt.model.components import ConfigComponent
@@ -54,6 +56,54 @@ class MDUComponent(ConfigComponent):
                 # model config is read if in read-mode and it exists
                 # default config if in write-mode
                 self.read()
+
+    def get_model_time(self):
+        """
+        Return (refdate, tstart, tstop) tuple.
+
+        It is parsed from model startdatetime/stopdatetime, or from the refdate/tunit/
+        tstart/tstop if not available.
+        """
+
+        def parse_time_to_datetime(key: str) -> dt.datetime:
+            date_str = str(self.get_value(key))
+            # validate datetime strings (from hydrolib.core)
+            date_str = validate_datetime_string(field_value=date_str, field=key)
+            # expected format is yyyymmddhhmmss, but hhmmss may be omitted
+            # (default:000000).
+            if len(date_str) == 8:
+                date_str += "000000"
+            date_dt = dt.datetime.strptime(date_str, "%Y%m%d%H%M%S")
+            return date_dt
+
+        startdatetime_str = self.get_value("time.startdatetime", "")
+        stopdatetime_str = self.get_value("time.stopdatetime", "")
+        if startdatetime_str == "" or stopdatetime_str == "":
+            logger.debug("get_model_time(): fallback to refdate/tstart/tstop")
+            refdate = parse_time_to_datetime(key="time.refdate")
+            tunit = self.get_value("time.tunit")
+            tstart = float(self.get_value("time.tstart"))
+            tstop = float(self.get_value("time.tstop"))
+            if tunit.lower() == "s":
+                tstart = refdate + dt.timedelta(seconds=tstart)
+                tstop = refdate + dt.timedelta(seconds=tstop)
+            elif tunit.lower() == "m":
+                tstart = refdate + dt.timedelta(minutes=tstart)
+                tstop = refdate + dt.timedelta(minutes=tstop)
+            elif tunit.lower() == "h":
+                tstart = refdate + dt.timedelta(hours=tstart)
+                tstop = refdate + dt.timedelta(hours=tstop)
+            elif tunit.lower() == "d":
+                tstart = refdate + dt.timedelta(days=tstart)
+                tstop = refdate + dt.timedelta(days=tstop)
+            else:
+                raise ValueError(f"tunit='{tunit}' not supported by get_model_time()")
+        else:
+            logger.debug("get_model_time(): from startdatetime/stopdatetime")
+            tstart = parse_time_to_datetime(key="time.startdatetime")
+            tstop = parse_time_to_datetime(key="time.stopdatetime")
+
+        return tstart, tstop
 
     @hydromt_step
     def read(self) -> None: